feat: console용/Web용 Command통합(인터페이스에 서로 다른 함수형인터페이스 인자를 가진 메서드를 한쪽에 추가)
