+++ src/main/java/chess/domain/command/End.java
@@ -1,6 +1,8 @@
 package chess.domain.command;
 
 import chess.domain.chessgame.ChessGame;
+import java.util.Map;
+import java.util.function.Supplier;
 
 public final class End implements CommandGenerator {
     @Override
@@ -19,5 +21,20 @@ public final class End implements CommandGenerator {
         printBoardInfoToState.run();
 
         chessGame.ready();
+    }
+
+    @Override
+    public Map<String, Object> execute(final String command,
+                                       final ChessGame chessGame,
+                                       final Supplier<Map<String, Object>> returnModelToState) {
+        if (!chessGame.isRunning()) {
+            chessGame.gameSwitchOff();
+        }
+
+        chessGame.end();
+        final Map<String, Object> model = returnModelToState.get();
+
+        chessGame.ready();
+        return model;
     }
 }
