+++ src/main/java/chess/domain/command/GameCommand.java
@@ -2,6 +2,7 @@ package chess.domain.command;
 
 import chess.domain.chessgame.ChessGame;
 import java.util.Arrays;
+import java.util.Map;
 import java.util.function.Supplier;
 import java.util.regex.Pattern;
 
@@ -18,7 +19,8 @@ public enum GameCommand {
     private final String command;
     private final Supplier<CommandGenerator> commandGenerator;
 
-    GameCommand(String command, Supplier<CommandGenerator> commandGenerator) {
+    GameCommand(String command,
+                Supplier<CommandGenerator> commandGenerator) {
         this.command = command;
         this.commandGenerator = commandGenerator;
     }
@@ -30,8 +32,17 @@ public enum GameCommand {
             .orElseThrow(() -> new IllegalArgumentException(WRONG_COMMAND_MESSAGE));
     }
 
-    public void execute(final String command, final ChessGame chessGame, final Runnable returnModelToState) {
+    public void execute(final String command,
+                        final ChessGame chessGame,
+                        final Runnable returnModelToState) {
         final CommandGenerator gameCommand = commandGenerator.get();
         gameCommand.execute(command, chessGame, returnModelToState);
+    }
+
+    public Map<String, Object> execute(final String command,
+                                       final ChessGame chessGame,
+                                       final Supplier<Map<String, Object>> returnModelToState) {
+        final CommandGenerator webGameCommand = commandGenerator.get();
+        return webGameCommand.execute(command, chessGame, returnModelToState);
     }
 }
