+++ src/main/java/chess/domain/command/Status.java
@@ -1,8 +1,11 @@
 package chess.domain.command;
 
 import chess.domain.chessgame.ChessGame;
+import java.util.Map;
+import java.util.function.Supplier;
 
 public final class Status implements CommandGenerator {
+
     @Override
     public void execute(final String command,
                         final ChessGame chessGame,
@@ -12,5 +15,14 @@ public final class Status implements CommandGenerator {
         printBoardInfoToState.run();
 
         chessGame.run();
+    }
+
+    @Override
+    public Map<String, Object> execute(final String command,
+                                       final ChessGame chessGame,
+                                       final Supplier<Map<String, Object>> returnModelToState) {
+        chessGame.status();
+        chessGame.run();
+        return null;
     }
 }
