--- src/main/java/chess/web/commandweb/Status.java
@@ -1,16 +1,0 @@
-package chess.web.commandweb;
-
-import chess.domain.chessgame.ChessGame;
-import java.util.Map;
-import java.util.function.Supplier;
-
-public final class Status implements WebCommandGenerator {
-    @Override
-    public Map<String, Object> execute(final String command,
-                                       final ChessGame chessGame,
-                                       final Supplier<Map<String, Object>> returnModelToState) {
-        chessGame.status();
-        chessGame.run();
-        return null;
-    }
-}
