--- src/main/java/chess/web/commandweb/WebCommandGenerator.java
@@ -1,11 +1,0 @@
-package chess.web.commandweb;
-
-import chess.domain.chessgame.ChessGame;
-import java.util.Map;
-import java.util.function.Supplier;
-
-public interface WebCommandGenerator {
-    Map<String, Object> execute(final String command,
-                                final ChessGame chessGame,
-                                final Supplier<Map<String, Object>> returnModelToState);
-}
